Remove old VC, move new to corresponding dirs
Modules names are still outdated
