+++ lib/membrane/video_compositor.ex
@@ -1,4 +1,4 @@
-defmodule Membrane.VideoCompositor do
+defmodule Membrane.VideoCompositor.MultipleInputs.VideoCompositor do
   @moduledoc """
   The element responsible for placing the first received frame
   above the other and sending forward buffer with
@@ -6,13 +6,13 @@ defmodule Membrane.VideoCompositor do
   """
 
   use Membrane.Filter
-  alias Membrane.Buffer
   alias Membrane.RawVideo
-  alias Membrane.VideoCompositor.Implementations
+  alias Membrane.VideoCompositor.MultipleInputs.VideoCompositor.Implementations
+  alias Membrane.VideoCompositor.MultipleInputs.VideoCompositor.Track
 
   def_options implementation: [
                 type: :atom,
-                spec: Implementations.implementation_t(),
+                spec: Implementations.implementation_t() | {:mock, module()},
                 description: "Implementation of video composer."
               ],
               caps: [
@@ -20,91 +20,256 @@ defmodule Membrane.VideoCompositor do
                 description: "Struct with video width, height, framerate and pixel format."
               ]
 
-  def_input_pad(:first_input,
+  def_input_pad :input,
     demand_unit: :buffers,
+    availability: :on_request,
     demand_mode: :auto,
-    caps: {RawVideo, pixel_format: :I420}
-  )
+    caps: {RawVideo, pixel_format: :I420},
+    options: [
+      position: [
+        type: :tuple,
+        spec: {integer(), integer()},
+        description:
+          "Initial position of the video on the screen, given in the pixels, relative to the upper left corner of the screen",
+        default: {0, 0}
+      ]
+    ]
 
-  def_input_pad(:second_input,
+  def_output_pad :output,
     demand_unit: :buffers,
     demand_mode: :auto,
     caps: {RawVideo, pixel_format: :I420}
-  )
-
-  def_output_pad(:output,
-    demand_unit: :buffers,
-    demand_mode: :auto,
-    caps: {RawVideo, pixel_format: :I420}
-  )
 
   @impl true
   def handle_init(options) do
-    {:ok, compositor_module} = Implementations.get_implementation_module(options.implementation)
+    compositor_module = determine_compositor_module(options.implementation)
 
     {:ok, internal_state} = compositor_module.init(options.caps)
 
     state = %{
-      pads: %{first_input: :queue.new(), second_input: :queue.new()},
-      streams_state: %{first_input: :playing, second_input: :playing},
+      ids_to_tracks: %{},
       caps: options.caps,
       compositor_module: compositor_module,
-      internal_state: internal_state
+      internal_state: internal_state,
+      pads_to_ids: %{},
+      new_pad_id: 0
     }
 
     {:ok, state}
   end
 
   @impl true
-  def handle_process(pad, buffer, _context, %{pads: pads} = state) do
-    updated_queue = Map.get(pads, pad)
-    updated_queue = :queue.in(buffer, updated_queue)
-    pads = Map.replace!(pads, pad, updated_queue)
-    state = %{state | pads: pads}
-
-    case {:queue.out(state.pads.first_input), :queue.out(state.pads.second_input)} do
-      {{{:value, first_frame_buffer}, rest_of_first_queue},
-       {{:value, second_frame_buffer}, rest_of_second_queue}} ->
-        frames_binaries = %{
-          first: first_frame_buffer.payload,
-          second: second_frame_buffer.payload
-        }
-
-        {{:ok, merged_frame_binary}, internal_state} =
-          state.compositor_module.merge_frames(frames_binaries, state.internal_state)
-
-        merged_image_buffer = %Buffer{first_frame_buffer | payload: merged_frame_binary}
-        pads = %{first_input: rest_of_first_queue, second_input: rest_of_second_queue}
-        state = %{state | pads: pads, internal_state: internal_state}
-        {{:ok, buffer: {:output, merged_image_buffer}}, state}
-
-      _one_of_queues_is_empty ->
-        {:ok, state}
-    end
+  def handle_prepared_to_playing(_ctx, state) do
+    {{:ok, caps: {:output, state.caps}}, state}
   end
 
   @impl true
-  def handle_caps(:first_input, %RawVideo{} = caps, _context, state) do
-    caps = %{caps | height: caps.height * 2}
-    {{:ok, caps: {:output, caps}}, state}
+  def handle_pad_added(pad, context, state) do
+    position = context.options.position
+
+    state = register_track(state, pad, position)
+    {:ok, state}
+  end
+
+  defp register_track(state, pad, position) do
+    new_id = state.new_pad_id
+
+    state = %{
+      state
+      | pads_to_ids: Map.put(state.pads_to_ids, pad, new_id),
+        new_pad_id: new_id + 1
+    }
+
+    %{state | ids_to_tracks: Map.put(state.ids_to_tracks, new_id, %Track{position: position})}
   end
 
   @impl true
-  def handle_caps(:second_input, %RawVideo{} = _caps, _context, state) do
+  def handle_caps(pad, caps, _context, state) do
+    %{
+      pads_to_ids: pads_to_ids,
+      internal_state: internal_state,
+      ids_to_tracks: ids_to_tracks
+    } = state
+
+    id = Map.get(pads_to_ids, pad)
+
+    position = Map.get(ids_to_tracks, id).position
+    {:ok, internal_state} = state.compositor_module.add_video(id, caps, position, internal_state)
+
+    state = %{state | internal_state: internal_state}
     {:ok, state}
   end
 
   @impl true
-  def handle_end_of_stream(pad, _context, %{streams_state: streams_state} = state) do
-    streams_state = Map.put(streams_state, pad, :end_of_the_stream)
-    state = %{state | streams_state: streams_state}
+  def handle_process(
+        pad,
+        buffer,
+        _context,
+        state
+      ) do
+    %{
+      ids_to_tracks: ids_to_tracks,
+      pads_to_ids: pads_to_ids
+    } = state
 
-    case {streams_state.first_input, streams_state.second_input} do
-      {:end_of_the_stream, :end_of_the_stream} ->
-        {{:ok, end_of_stream: :output, notify: {:end_of_stream, pad}}, state}
+    id = Map.get(pads_to_ids, pad)
 
-      _one_streams_has_not_ended ->
+    ids_to_tracks = push_frame(ids_to_tracks, id, buffer)
+
+    state = %{state | ids_to_tracks: ids_to_tracks}
+
+    case merge_frames(state) do
+      {{:merged, buffers}, state} -> {{:ok, buffer: {:output, buffers}}, state}
+      {{:empty, _empty}, state} -> {:ok, state}
+    end
+  end
+
+  defp merge_frames(state) do
+    %{
+      ids_to_tracks: ids_to_tracks,
+      internal_state: internal_state
+    } = state
+
+    if all_tracks_ready_to_merge?(ids_to_tracks) do
+      ids_to_frames = get_ids_to_frames(ids_to_tracks)
+
+      {{:ok, merged_frame_binary}, internal_state} =
+        state.compositor_module.merge_frames(ids_to_frames, internal_state)
+
+      ids_to_tracks = pop_frames(ids_to_tracks)
+
+      state = %{
+        state
+        | ids_to_tracks: ids_to_tracks,
+          internal_state: internal_state
+      }
+
+      state = remove_finished_tracks(state)
+
+      {{_status, tail}, state} = merge_frames(state)
+      merged = [%Membrane.Buffer{payload: merged_frame_binary}] ++ tail
+
+      {{:merged, merged}, state}
+    else
+      {{:empty, []}, state}
+    end
+  end
+
+  defp push_frame(ids_to_tracks, id, frame) do
+    Map.update!(ids_to_tracks, id, fn track ->
+      Track.push_frame(track, frame)
+    end)
+  end
+
+  defp all_tracks_ready_to_merge?(ids_to_tracks) when map_size(ids_to_tracks) == 0 do
+    false
+  end
+
+  defp all_tracks_ready_to_merge?(ids_to_tracks) do
+    ids_to_tracks
+    |> Map.values()
+    |> Enum.all?(&Track.has_frame?/1)
+  end
+
+  defp get_ids_to_frames(ids_to_tracks) do
+    ids_to_tracks
+    |> Enum.map(fn {id, %Track{} = track} -> {id, Track.first_frame(track)} end)
+    |> Enum.sort()
+  end
+
+  defp pop_frames(ids_to_tracks) do
+    ids_to_tracks
+    |> Enum.map(fn {id, %Track{} = track} ->
+      {id, Track.pop_frame(track)}
+    end)
+    |> Map.new()
+  end
+
+  defp remove_finished_tracks(state) do
+    state.ids_to_tracks
+    |> Enum.reduce(
+      state,
+      fn {id, %Track{} = track}, state ->
+        if Track.finished?(track) do
+          remove_track(state, id)
+        else
+          state
+        end
+      end
+    )
+  end
+
+  defp track_finished?(ids_to_tracks, id) do
+    track = Map.get(ids_to_tracks, id)
+    Track.finished?(track)
+  end
+
+  defp remove_track(state, id) do
+    %{ids_to_tracks: ids_to_tracks, internal_state: internal_state} = state
+    ids_to_tracks = Map.delete(ids_to_tracks, id)
+    {:ok, internal_state} = state.compositor_module.remove_video(id, internal_state)
+    %{state | ids_to_tracks: ids_to_tracks, internal_state: internal_state}
+  end
+
+  defp update_track_status(ids_to_tracks, id, status) do
+    ids_to_tracks
+    |> Map.update!(id, fn track -> %Track{track | status: status} end)
+  end
+
+  defp tracks_status(ids_to_tracks) do
+    if Map.values(ids_to_tracks) |> Enum.all?(&Track.finished?/1) do
+      :all_finished
+    else
+      :still_ongoing
+    end
+  end
+
+  @impl true
+  def handle_end_of_stream(
+        pad,
+        _context,
+        state
+      ) do
+    %{ids_to_tracks: ids_to_tracks, pads_to_ids: pads_to_ids} = state
+    id = Map.get(pads_to_ids, pad)
+
+    ids_to_tracks = ids_to_tracks |> update_track_status(id, :end_of_stream)
+    state = %{state | ids_to_tracks: ids_to_tracks}
+
+    state =
+      if track_finished?(ids_to_tracks, id) do
+        remove_track(state, id)
+      else
+        state
+      end
+
+    {{buffers_status, buffers}, state} = merge_frames(state)
+
+    case {buffers_status, tracks_status(state.ids_to_tracks)} do
+      {:empty, :still_ongoing} ->
         {:ok, state}
+
+      {:empty, :all_finished} ->
+        {{:ok, end_of_stream: :output}, state}
+
+      {:merged, :still_ongoing} ->
+        {{:ok, buffer: {:output, buffers}}, state}
+
+      {:merged, :all_finished} ->
+        {{:ok, buffer: {:output, buffers}, end_of_stream: :output}, state}
+    end
+  end
+
+  defp determine_compositor_module(implementation) do
+    case implementation do
+      {:mock, module} ->
+        module
+
+      implementation ->
+        case Implementations.get_implementation_module(implementation) do
+          {:ok, module} -> module
+          {:error, error} -> raise error
+        end
     end
   end
 end
