+++ lib/membrane/video_compositor/implementation.ex
@@ -1,6 +1,6 @@
-defmodule Membrane.VideoCompositor.Implementations do
+defmodule Membrane.VideoCompositor.MultipleInputs.VideoCompositor.Implementations do
   @moduledoc """
-  A module describing video compositor implementation type and implementing
+  A module describing multiple input video compositor implementation type and implementing
   functions related with implementation format.
   """
 
@@ -11,35 +11,30 @@ defmodule Membrane.VideoCompositor.Implementations do
   def get_implementation_module(implementation) do
     case implementation do
       :ffmpeg ->
-        {:ok, Membrane.VideoCompositor.Implementations.FFmpeg}
+        raise ":ffmpeg is not implemented yet"
 
       :opengl_cpp ->
-        {:ok, Membrane.VideoCompositor.Implementations.OpenGL.Cpp}
+        raise ":opengl_cpp is not implemented yet"
 
       :opengl_rust ->
-        {:ok, Membrane.VideoCompositor.Implementations.OpenGL.Rust}
+        {:ok, Membrane.VideoCompositor.MultipleInputs.OpenGL.Rust}
 
       :nx ->
-        {:ok, Membrane.VideoCompositor.Nx}
+        raise ":nx is not implemented yet"
 
       _other ->
         {:error, "Format not supported"}
     end
   end
 
-  @spec get_test_implementations() :: list()
-  def get_test_implementations() do
-    test_implementations = [:ffmpeg, :nx]
-    test_implementations
-  end
-
-  @spec get_all_implementations() :: list()
+  @spec get_all_implementations() :: list(implementation_t)
   def get_all_implementations() do
-    [:ffmpeg, :opengl_cpp, :opengl_rust, :nx]
+    [:opengl_rust]
   end
 
   @spec get_implementation_atom_from_string(String.t()) :: implementation_t()
-  def get_implementation_atom_from_string(implementation_string) do
+  def get_implementation_atom_from_string(implementation_string)
+      when is_binary(implementation_string) do
     case implementation_string do
       "ffmpeg" -> :ffmpeg
       "opengl_cpp" -> :opengl_cpp
