--- lib/membrane/video_compositor/implementation/opengl/native/cpp.ex
@@ -1,4 +1,0 @@
-defmodule Membrane.VideoCompositor.Implementations.OpenGL.Native.Cpp do
-  @moduledoc false
-  use Unifex.Loader
-end
