+++ lib/membrane/video_compositor/track.ex
